refactor: single cmd 일 때 fork 없이 부모에서 실행하는 것 완료
refactoring 완료 후, 2차원 배열 free 하는 함수 libft 에 추가
