+++ kipark_jayoon_minishell/includes/execution.h
@@ -6,7 +6,7 @@
 /*   By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/21 20:37:08 by jayoon            #+#    #+#             */
-/*   Updated: 2022/09/24 16:16:05 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/09/24 21:40:33 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 
 # include "env.h"
 # include "parser.h"
-# include "env.h"
 
 typedef struct s_args_execve
 {
@@ -23,6 +22,13 @@ typedef struct s_args_execve
 	char	**argv;
 }	t_args_execve;
 
+// main
 void	execute_cmd(t_parsing_list *l_parsing, t_env *l_env);
+void	process_execve_args(t_parsing_list *l_parsing, t_args_execve *p_args,
+			t_env *l_env);
+char	**init_curr_envp(t_env *l_env);
+
+// utils
+int		is_single_cmd(t_parsing_list *next);
 
 #endif
