+++ kipark_jayoon_minishell/libft/Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+         #
+#    By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/11/20 18:08:18 by jayoon            #+#    #+#              #
-#    Updated: 2022/09/24 15:48:24 by kipark           ###   ########seoul.kr   #
+#    Updated: 2022/09/24 21:49:39 by jayoon           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -55,6 +55,7 @@ SRCS =		ft_isalnum.c			\
 			get_next_line.c			\
 			get_next_line_utils.c	\
 			ft_safe_free.c			\
+			ft_safe_free_two_dimentions_arr.c \
 			ft_safe_malloc.c		\
 			ft_check_error.c		\
 			ft_strchr_index.c		\
