+++ kipark_jayoon_minishell/srcs/execution/is_single_cmd.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   is_single_cmd.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/24 21:09:25 by jayoon            #+#    #+#             */
+/*   Updated: 2022/09/24 21:10:00 by jayoon           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parser.h"
+#include <stdlib.h>
+
+int	is_single_cmd(t_parsing_list *next)
+{
+	if (next == NULL)
+		return (1);
+	return (0);
+}
